code refactoring from todo

+++ app/common/tables/base.py
@@ -1,9 +1,9 @@
-from sqlalchemy import Integer
+from sqlalchemy import BigInteger
 from sqlalchemy.orm import DeclarativeBase, declared_attr, mapped_column
 
 
 class BaseTable(DeclarativeBase):
-    id = mapped_column(Integer, primary_key=True, nullable=False, autoincrement=True)
+    id = mapped_column(BigInteger, primary_key=True, nullable=False, autoincrement=True)
 
     @declared_attr
     def __tablename__(cls) -> str:
